+++ src/structs.rs
@@ -22,7 +22,7 @@ pub struct ContainerForSend {
 
 #[derive(Clone, Debug, PartialEq, Eq, serde::Deserialize, serde::Serialize, Hash)]
 pub struct WindowForSend {
-    pub id: u64,
+    pub id: u16,
     pub icon: Icon,
     pub accent_color: u32,
 }
